implement createteam and update endpoints for new internal billing
